new feature added if users table is empty in the manage-users page then table does not show instead an error message with 'no users found' appear

+++ admin/manage-users.php
@@ -98,30 +98,36 @@ $users = mysqli_query($connection, $query);
             </aside>
             <main>
                 <h2>Manage Users</h2>
-                <table>
-                    <thead>
-                        <tr>
-                            <th>Name</th>
-                            <th>Username</th>
-                            <th>Edit</th>
-                            <th>Delete</th>
-                            <th>Admin</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php while($user = mysqli_fetch_assoc($users)) : ?>
-                        <tr>
-                            <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
-                            <td><?= "{$user['username']}" ?></td>
-                            <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
-                            <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
-                            <td>
-                                <?= $user['is_admin'] ? "yes" : "no" ?>
-                            </td>
-                        </tr> <!--End item 1-->
-                        <?php endwhile ?>
-                    </tbody>
-                </table>
+                <?php if(mysqli_num_rows($users) > 0) : ?>
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>Name</th>
+                                <th>Username</th>
+                                <th>Edit</th>
+                                <th>Delete</th>
+                                <th>Admin</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php while($user = mysqli_fetch_assoc($users)) : ?>
+                            <tr>
+                                <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
+                                <td><?= "{$user['username']}" ?></td>
+                                <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
+                                <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
+                                <td>
+                                    <?= $user['is_admin'] ? "yes" : "no" ?>
+                                </td>
+                            </tr> <!--End item 1-->
+                            <?php endwhile ?>
+                        </tbody>
+                    </table>
+                <?php else : ?>
+                    <div class="alert__message error">
+                        <p><?= "No users found" ?></p>
+                    </div>
+                <?php endif ?>
             </main>
         </div>
     </section>
